requirements.txt
flask
gymnasium
numpy
pillow
tensorflow
keras